feat(batch_processor): add batch processor module
Add batch_processor module to the project. This module includes a Flask application that processes uploaded tasks in the background using Celery and Redis. The module consists of the following files:

- `README.md`: Added a README file with environment variable configurations for the Celery broker and result backend.
- `app.py`: Added the main Flask application file that configures Celery, SQLAlchemy, and defines functions for processing uploaded tasks and scheduling the task processing.
- `docker-compose.yml`: Added a Docker Compose file to define the services for the batch processor, including the worker service for running Celery, and the Redis service for the broker and result backend.
- `Dockerfile`: Added a Dockerfile to build the batch processor image, which installs the necessary dependencies and sets up the application.
- `requirements.txt`: Added a requirements file with the necessary Python dependencies for the batch processor.

The batch processor module allows for efficient processing of uploaded

+++ api/resources/tasks.py
@@ -44,7 +44,7 @@ class Tasks(Resource):
         unique_filename = str(uuid.uuid4()) + file_extension
 
         # Usa el nombre único con extensión al guardar
-        save_path = os.path.join("/Users/felixernestoorduzgrimaldo/Documents/Estudio/MISO/test_temp/", unique_filename)
+        save_path = os.path.join(os.environ.get('SAVE_PATH', '/file_conversor/uploaded/'), unique_filename)
 
         # Guarda el archivo en la ruta especificada
         uploaded_file.save(save_path)
@@ -125,6 +125,9 @@ class Tasks(Resource):
 
         if task is None:
             return {"message": "Tarea no encontrada"}, 404
+
+        if task.status is not 'processed':
+            return {"message": "Tarea no esta en estado processed"}, 400
 
         # Elimina la tarea de la base de datos
         db.session.delete(task)
